+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/tsf/adaptor/PolarisAdaptorTsfConfigExtensionLayer.java
@@ -95,9 +95,7 @@ public class PolarisAdaptorTsfConfigExtensionLayer implements PolarisConfigCusto
 	}
 
 	/**
-	 * @see PolarisConfigCustomExtensionLayer#initConfigFiles(CompositePropertySource,
-	 *        PolarisPropertySourceManager,
-	 *        ConfigFileService)
+	 * @see PolarisConfigCustomExtensionLayer#initConfigFiles(Environment , CompositePropertySource , ConfigFileService )
 	 */
 	@Override
 	public void initConfigFiles(Environment environment, CompositePropertySource compositePropertySource,