+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/tsf/env/TsfCoreEnvironmentPostProcessor.java
@@ -59,7 +59,7 @@ public final class TsfCoreEnvironmentPostProcessor implements EnvironmentPostPro
 		String tsfAppId = environment.getProperty("tsf_app_id");
 		if (StringUtils.isNotBlank(tsfAppId)) {
 			Map<String, Object> defaultProperties = new HashMap<>();
-			//默认开启polaris
+			//enable polaris as default
 			defaultProperties.put("spring.cloud.polaris.enabled", "true");
 			// tsf_application_id
 			String tsfApplicationId = environment.getProperty("tsf_application_id");